pyproject.toml
[tool.poetry]
name = "pyproject"
version = "0.1.0"
description = ""
authors = ["poetry install"]
license = "poetry install"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
requests = "^2.31.0"
python-dotenv = "^1.0.1"


[tool.poetry.group.lint.dependencies]
black = "^24.3.0"
flake8 = "^7.0.0"
isort = "^5.13.2"
mypy = "^1.9.0"


[tool.poetry.group.dev.dependencies]
pytest = "^8.1.1"
pytest-cov = "^5.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.mypy]
disallow_untyped_defs = true
no_implicit_optional = true
warn_return_any = true
exclude = 'venv'

[tool.isort]
line_length = 120
profile = "black"
multi_line_output = 3

[tool.black]
line-length = 120
exclude = '''
(
  /(
      \.eggs
    | \.git
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
)
)
'''
